moved button-footer use the provided getUserInfo method
