+++ coldstartapp/vue-app/src/store/modules/mutation-types.js
@@ -1,5 +1,3 @@
-const GET_CATALOG = 'GET_CATALOG';
+export const GET_CATALOG = 'GET_CATALOG';
 export const GET_ORDERS = 'GET_ORDERS';
 export const POST_ORDERS = 'POST_ORDERS';
-
-export { GET_CATALOG as default };
